feat: refactor calculator module by organizing types, utils, and configuration
